refactor(api): fix Clean Architecture layer violations (MGG-164)
- Move AuthEventType enum from Infrastructure to Common layer
- Replace DbUpdateException catches in controllers with DuplicateEntityException
- Move refresh token lookup from EF Core query in AuthController to IUserService
- Remove all EF Core and Infrastructure.Entities.Audit refs from Presentation
