+++ src/Common/GlobalEnums.cs
@@ -10,3 +10,16 @@ public enum PricingMode
 	Quantity,
 	Flat
 }
+
+public enum AuthEventType
+{
+	Login,
+	LoginFailed,
+	Logout,
+	ApiKeyUsed,
+	ApiKeyCreated,
+	ApiKeyRevoked,
+	PasswordChanged,
+	UserRegistered,
+	AccountDisabled,
+}
