--- src/Infrastructure/Entities/Audit/AuthEventType.cs
@@ -1,14 +1,0 @@
-namespace Infrastructure.Entities.Audit;
-
-public enum AuthEventType
-{
-	Login,
-	LoginFailed,
-	Logout,
-	ApiKeyUsed,
-	ApiKeyCreated,
-	ApiKeyRevoked,
-	PasswordChanged,
-	UserRegistered,
-	AccountDisabled,
-}
