+++ src/Infrastructure/Services/AuthAuditService.cs
@@ -1,4 +1,5 @@
 using Application.Interfaces.Services;
+using Common;
 using Infrastructure.Entities.Audit;
 using Microsoft.EntityFrameworkCore;
