+++ src/Presentation/API/Controllers/ApiKeyController.cs
@@ -1,6 +1,6 @@
 using API.Generated.Dtos;
 using Application.Interfaces.Services;
-using Infrastructure.Entities.Audit;
+using Common;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Security.Claims;
