+++ src/Presentation/API/Controllers/Core/SubcategoriesController.cs
@@ -4,12 +4,12 @@ using Application.Commands.Subcategory.Create;
 using Application.Commands.Subcategory.Delete;
 using Application.Commands.Subcategory.Restore;
 using Application.Commands.Subcategory.Update;
+using Application.Exceptions;
 using Application.Queries.Core.Subcategory;
 using Domain.Core;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
 
 namespace API.Controllers.Core;
 
@@ -148,10 +148,10 @@ public class SubcategoriesController(IMediator mediator, SubcategoryMapper mappe
 			List<Subcategory> subcategories = await mediator.Send(command);
 			return Ok(mapper.ToResponse(subcategories[0]));
 		}
-		catch (DbUpdateException ex) when (ex.InnerException?.Message.Contains("unique", StringComparison.OrdinalIgnoreCase) == true)
+		catch (DuplicateEntityException ex)
 		{
 			logger.LogWarning(ex, "Duplicate subcategory name in {Method}", nameof(CreateSubcategory));
-			return Conflict("A subcategory with this name already exists in this category.");
+			return Conflict(ex.Message);
 		}
 		catch (Exception ex)
 		{
@@ -174,10 +174,10 @@ public class SubcategoriesController(IMediator mediator, SubcategoryMapper mappe
 			List<Subcategory> subcategories = await mediator.Send(command);
 			return Ok(subcategories.Select(mapper.ToResponse).ToList());
 		}
-		catch (DbUpdateException ex) when (ex.InnerException?.Message.Contains("unique", StringComparison.OrdinalIgnoreCase) == true)
+		catch (DuplicateEntityException ex)
 		{
 			logger.LogWarning(ex, "Duplicate subcategory name in {Method}", nameof(CreateSubcategories));
-			return Conflict("A subcategory with this name already exists in this category.");
+			return Conflict(ex.Message);
 		}
 		catch (Exception ex)
 		{
