+++ src/Presentation/API/Controllers/UsersController.cs
@@ -1,7 +1,7 @@
 using API.Generated.Dtos;
 using Application.Interfaces.Services;
+using Common;
 using Infrastructure.Entities;
-using Infrastructure.Entities.Audit;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
