+++ tests/Infrastructure.Tests/Services/AuthAuditServiceTests.cs
@@ -1,4 +1,5 @@
 using Application.Interfaces.Services;
+using Common;
 using FluentAssertions;
 using Infrastructure.Entities.Audit;
 using Infrastructure.Services;
